refactor: split product and import stacks
