+++ cdk/app.ts
@@ -1,13 +1,20 @@
+import 'dotenv/config';
 import * as cdk from 'aws-cdk-lib';
 import { ProductsApiStack } from './lib/productsApiStack';
 import { SnsStack } from './lib/snsStack';
+import { SqsStack } from './lib/sqsStack';
+import { ImportApiStack } from './lib/importApiStack';
 
 const app = new cdk.App();
 
-const productCreationSnsStack = new SnsStack(app, 'ProductCreationSnsStack');
+const { productCreationTopic } = new SnsStack(app, 'ProductCreationSnsStack');
+const { productCreationQueue } = new SqsStack(app, 'ProductCreationQueueStack');
 
 new ProductsApiStack(app, 'ProductsApiStack', {
-  productCreationTopic: productCreationSnsStack.productCreationTopic,
+  productCreationTopic,
+  productCreationQueue,
 });
+
+new ImportApiStack(app, 'ImportApiStack', { productCreationQueue });
 
 app.synth();
