+++ cdk/lib/snsStack.ts
@@ -1,4 +1,3 @@
-import 'dotenv/config';
 import * as cdk from 'aws-cdk-lib';
 import * as sns from 'aws-cdk-lib/aws-sns';
 import * as subscriptions from 'aws-cdk-lib/aws-sns-subscriptions';
