+++ package.json
@@ -6,9 +6,11 @@
   "scripts": {
     "lint": "eslint .",
     "tsc": "tsc",
-    "cdk:bootstrap": "npx cdk bootstrap",
-    "cdk:deploy:all": "npm run lint && npm run tsc && npx cdk deploy --all",
-    "cdk:deploy:products": "npm run lint && npm run tsc && npx cdk deploy ProductsApiStack",
+    "prebuild": "npm run lint && npm run tsc",
+    "bootstrap": "npx cdk bootstrap",
+    "deploy:all": "npm run prebuild && npx cdk deploy --all",
+    "deploy:products": "npm run prebuild && npx cdk deploy ProductsApiStack",
+    "deploy:import": "npm run prebuild && npx cdk deploy ImportApiStack",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
